MDEV-4991: Minor fixes for issues seen in Buildbot

+++ sql/log.cc
@@ -3867,7 +3867,7 @@ bool MYSQL_BIN_LOG::open(const char *log_name,
         if (gtid_index)
           delete gtid_index;
         my_off_t offset= my_b_tell(&log_file);
-        gtid_index= new Gtid_index_writer(log_file_name, offset,
+        gtid_index= new Gtid_index_writer(log_file_name, (uint32)offset,
                                           &rpl_global_gtid_binlog_state);
         if (!gtid_index)
           sql_print_information("Could not create GTID index for binlog "
@@ -7177,7 +7177,7 @@ err:
         if (likely(gtid_index))
         {
           rpl_gtid gtid= thd->get_last_commit_gtid();
-          gtid_index->process_gtid(offset, &gtid);
+          gtid_index->process_gtid((uint32)offset, &gtid);
         }
 
         if ((error= flush_and_sync(&synced)))
@@ -8720,7 +8720,7 @@ MYSQL_BIN_LOG::trx_group_commit_leader(group_commit_entry *leader)
       if (likely(gtid_index))
       {
         rpl_gtid gtid= current->thd->get_last_commit_gtid();
-        gtid_index->process_gtid(commit_offset, &gtid);
+        gtid_index->process_gtid((uint32)commit_offset, &gtid);
       }
       cache_mngr->last_commit_pos_offset= commit_offset;
       if ((cache_mngr->using_xa && cache_mngr->xa_xid) || current->need_unlog)
